Player Stats Implementation Started
I have the logic for stat modifiers and stat changing. The actual stats are not yet modifier. I also set the FPS of the scene to 60, and doing so, I had to modify some of the values in the dodging logic. I also created the Player class which will have the stats and handle most of the actions a player takes. This includes the Move() function in the PlayerMovement script. I commented out the Update() function in the PlayerMovement script and made the Player class call Move() instead. Let me know if there is an issue with these changes, and I will work on fixing them to better suit better ideas.