GameOver Screen, Win Screen, and Contact Damage Implemented
Contact Damage is set at a static 10 for now as we don't have other enemies yet. Going to 0 or below 0 health will trigger the game over screen with button not yet implemented. Touching the win condition, in this case the book, will make the win screen appear with buttons not yet implemented.